Symmetry sectors and fusion trees

Type hierarchy

Missing docstring.

Missing docstring for Sector. Check Documenter's build log for details.

Missing docstring.

Missing docstring for SectorValues. Check Documenter's build log for details.

Missing docstring.

Missing docstring for FusionStyle. Check Documenter's build log for details.

Missing docstring.

Missing docstring for BraidingStyle. Check Documenter's build log for details.

Missing docstring.

Missing docstring for AbstractIrrep. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Trivial. Check Documenter's build log for details.

Missing docstring.

Missing docstring for ZNIrrep. Check Documenter's build log for details.

Missing docstring.

Missing docstring for U1Irrep. Check Documenter's build log for details.

Missing docstring.

Missing docstring for SU2Irrep. Check Documenter's build log for details.

Missing docstring.

Missing docstring for CU1Irrep. Check Documenter's build log for details.

Missing docstring.

Missing docstring for ProductSector. Check Documenter's build log for details.

Missing docstring.

Missing docstring for FermionParity. Check Documenter's build log for details.

Missing docstring.

Missing docstring for FermionNumber. Check Documenter's build log for details.

Missing docstring.

Missing docstring for FermionSpin. Check Documenter's build log for details.

Missing docstring.

Missing docstring for FibonacciAnyon. Check Documenter's build log for details.

Missing docstring.

Missing docstring for IsingAnyon. Check Documenter's build log for details.

TensorKit.FusionTree — Type

struct FusionTree{I, N, M, L, T}

Represents a fusion tree of sectors of type I<:Sector, fusing (or splitting) N uncoupled sectors to a coupled sector. It actually represents a splitting tree, but fusion tree is a more common term.

Fields

uncoupled::NTuple{N,I}: the uncoupled sectors coming out of the splitting tree, before the possible 𝑍 isomorphism (see isdual).

coupled::I: the coupled sector.

isdual::NTuple{N,Bool}: indicates whether a 𝑍 isomorphism is present (true) or not (false) for each uncoupled sector.

innerlines::NTuple{M,I}: the labels of the M=max(0, N-2) inner lines of the splitting tree.

vertices::NTuple{L,T}: the L=max(0, N-1) labels of type T of the vertices of the splitting tree. If FusionStyle(I) isa MultiplicityFreeFusion, then T = Nothing.

Useful constants

Missing docstring.

Missing docstring for Irrep. Check Documenter's build log for details.

Methods for defining and characterizing Sector subtypes

Missing docstring.

Missing docstring for Base.one(::Sector). Check Documenter's build log for details.

Missing docstring.

Missing docstring for dual(::Sector). Check Documenter's build log for details.

Missing docstring.

Missing docstring for Nsymbol. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Fsymbol. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Rsymbol. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Bsymbol. Check Documenter's build log for details.

TensorKitSectors.dim — Method

dim(P::ProductSpace{S, N}, s::NTuple{N, sectortype(S)}) where {S<:ElementarySpace}
-> Int

Return the total degeneracy dimension corresponding to a tuple of sectors for each of the spaces in the tensor product, obtained as prod(dims(P, s))`.

Missing docstring.

Missing docstring for frobeniusschur. Check Documenter's build log for details.

Missing docstring.

Missing docstring for twist(::Sector). Check Documenter's build log for details.

Missing docstring.

Missing docstring for Base.isreal(::Type{<:Sector}). Check Documenter's build log for details.

Missing docstring.

Missing docstring for TensorKit.vertex_labeltype. Check Documenter's build log for details.

Missing docstring.

Missing docstring for TensorKit.vertex_ind2label. Check Documenter's build log for details.

Missing docstring.

Missing docstring for ⊠(::Sector, ::Sector). Check Documenter's build log for details.

TensorKit.fusiontrees — Method

fusiontrees(uncoupled::NTuple{N,I}[,
    coupled::I=one(I)[, isdual::NTuple{N,Bool}=ntuple(n -> false, length(uncoupled))]])
    where {N,I<:Sector} -> FusionTreeIterator{I,N}

Return an iterator over all fusion trees with a given coupled sector label coupled and uncoupled sector labels and isomorphisms uncoupled and isdual respectively.

Methods for manipulating fusion trees

For manipulating single fusion trees, the following internal methods are defined:

TensorKit.insertat — Function

insertat(f::FusionTree{I, N₁}, i::Int, f₂::FusionTree{I, N₂})
-> <:AbstractDict{<:FusionTree{I, N₁+N₂-1}, <:Number}

Attach a fusion tree f₂ to the uncoupled leg i of the fusion tree f₁ and bring it into a linear combination of fusion trees in standard form. This requires that f₂.coupled == f₁.uncoupled[i] and f₁.isdual[i] == false.

TensorKit.split — Function

split(f::FusionTree{I, N}, M::Int)
-> (::FusionTree{I, M}, ::FusionTree{I, N-M+1})

Split a fusion tree into two. The first tree has as uncoupled sectors the first M uncoupled sectors of the input tree f, whereas its coupled sector corresponds to the internal sector between uncoupled sectors M and M+1 of the original tree f. The second tree has as first uncoupled sector that same internal sector of f, followed by remaining N-M uncoupled sectors of f. It couples to the same sector as f. This operation is the inverse of insertat in the sense that if f₁, f₂ = split(t, M) ⇒ f == insertat(f₂, 1, f₁).

TensorKit.merge — Function

merge(f₁::FusionTree{I, N₁}, f₂::FusionTree{I, N₂}, c::I, μ = nothing)
-> <:AbstractDict{<:FusionTree{I, N₁+N₂}, <:Number}

Merge two fusion trees together to a linear combination of fusion trees whose uncoupled sectors are those of f₁ followed by those of f₂, and where the two coupled sectors of f₁ and f₂ are further fused to c. In case of FusionStyle(I) == GenericFusion(), also a degeneracy label μ for the fusion of the coupled sectors of f₁ and f₂ to c needs to be specified.

TensorKit.elementary_trace — Function

elementary_trace(f::FusionTree{I,N}, i) where {I<:Sector,N} -> <:AbstractDict{FusionTree{I,N-2}, <:Number}

Perform an elementary trace of neighbouring uncoupled indices i and i+1 on a fusion tree f, and returns the result as a dictionary of output trees and corresponding coefficients.

TensorKit.planar_trace — Method

planar_trace(f::FusionTree{I,N}, q1::IndexTuple{N₃}, q2::IndexTuple{N₃}) where {I<:Sector,N,N₃}
    -> <:AbstractDict{FusionTree{I,N-2*N₃}, <:Number}

Perform a planar trace of the uncoupled indices of the fusion tree f at q1 with those at q2, where q1[i] is connected to q2[i] for all i. The result is returned as a dictionary of output trees and corresponding coefficients.

TensorKit.artin_braid — Function

artin_braid(f::FusionTree, i; inv::Bool = false) -> <:AbstractDict{typeof(f), <:Number}

Perform an elementary braid (Artin generator) of neighbouring uncoupled indices i and i+1 on a fusion tree f, and returns the result as a dictionary of output trees and corresponding coefficients.

The keyword inv determines whether index i will braid above or below index i+1, i.e. applying artin_braid(f′, i; inv = true) to all the outputs f′ of artin_braid(f, i; inv = false) and collecting the results should yield a single fusion tree with non-zero coefficient, namely f with coefficient 1. This keyword has no effect if BraidingStyle(sectortype(f)) isa SymmetricBraiding.

TensorKit.braid — Method

braid(f::FusionTree{<:Sector, N}, levels::NTuple{N, Int}, p::NTuple{N, Int})
-> <:AbstractDict{typeof(t), <:Number}

Perform a braiding of the uncoupled indices of the fusion tree f and return the result as a <:AbstractDict of output trees and corresponding coefficients. The braiding is determined by specifying that the new sector at position k corresponds to the sector that was originally at the position i = p[k], and assigning to every index i of the original fusion tree a distinct level or depth levels[i]. This permutation is then decomposed into elementary swaps between neighbouring indices, where the swaps are applied as braids such that if i and j cross, $τ_{i,j}$ is applied if levels[i] < levels[j] and $τ_{j,i}^{-1}$ if levels[i] > levels[j]. This does not allow to encode the most general braid, but a general braid can be obtained by combining such operations.

TensorKit.permute — Method

permute(f::FusionTree, p::NTuple{N, Int}) -> <:AbstractDict{typeof(t), <:Number}

Perform a permutation of the uncoupled indices of the fusion tree f and returns the result as a <:AbstractDict of output trees and corresponding coefficients; this requires that BraidingStyle(sectortype(f)) isa SymmetricBraiding.

These can be composed to manipulate fusion-splitting tree pairs, for which the following methods are defined:

Missing docstring.

Missing docstring for bendright. Check Documenter's build log for details.

Missing docstring.

Missing docstring for bendleft. Check Documenter's build log for details.

Missing docstring.

Missing docstring for foldright. Check Documenter's build log for details.

Missing docstring.

Missing docstring for foldleft. Check Documenter's build log for details.

Missing docstring.

Missing docstring for cycleclockwise. Check Documenter's build log for details.

Missing docstring.

Missing docstring for cycleanticlockwise. Check Documenter's build log for details.

TensorKit.repartition — Function

repartition(f₁::FusionTree{I, N₁}, f₂::FusionTree{I, N₂}, N::Int) where {I, N₁, N₂}
-> <:AbstractDict{Tuple{FusionTree{I, N}, FusionTree{I, N₁+N₂-N}}, <:Number}

Input is a double fusion tree that describes the fusion of a set of incoming uncoupled sectors to a set of outgoing uncoupled sectors, represented using the individual trees of outgoing (f₁) and incoming sectors (f₂) respectively (with identical coupled sector f₁.coupled == f₂.coupled). Computes new trees and corresponding coefficients obtained from repartitioning the tree by bending incoming to outgoing sectors (or vice versa) in order to have N outgoing sectors.

repartition(tsrc::AbstractTensorMap{S}, N₁::Int, N₂::Int; copy::Bool=false) where {S}
    -> tdst::AbstractTensorMap{S,N₁,N₂}

Return tensor tdst obtained by repartitioning the indices of t. The codomain and domain of tdst correspond to the first N₁ and last N₂ spaces of t, respectively.

If copy=false, tdst might share data with tsrc whenever possible. Otherwise, a copy is always made.

To repartition into an existing destination, see repartition!.

Base.transpose — Method

transpose(f₁::FusionTree{I}, f₂::FusionTree{I},
        p1::NTuple{N₁, Int}, p2::NTuple{N₂, Int}) where {I, N₁, N₂}
-> <:AbstractDict{Tuple{FusionTree{I, N₁}, FusionTree{I, N₂}}, <:Number}

Input is a double fusion tree that describes the fusion of a set of incoming uncoupled sectors to a set of outgoing uncoupled sectors, represented using the individual trees of outgoing (t1) and incoming sectors (t2) respectively (with identical coupled sector t1.coupled == t2.coupled). Computes new trees and corresponding coefficients obtained from repartitioning and permuting the tree such that sectors p1 become outgoing and sectors p2 become incoming.

TensorKit.braid — Method

braid(f₁::FusionTree{I}, f₂::FusionTree{I},
        levels1::IndexTuple, levels2::IndexTuple,
        p1::IndexTuple{N₁}, p2::IndexTuple{N₂}) where {I<:Sector, N₁, N₂}
-> <:AbstractDict{Tuple{FusionTree{I, N₁}, FusionTree{I, N₂}}, <:Number}

Input is a fusion-splitting tree pair that describes the fusion of a set of incoming uncoupled sectors to a set of outgoing uncoupled sectors, represented using the splitting tree f₁ and fusion tree f₂, such that the incoming sectors f₂.uncoupled are fused to f₁.coupled == f₂.coupled and then to the outgoing sectors f₁.uncoupled. Compute new trees and corresponding coefficients obtained from repartitioning and braiding the tree such that sectors p1 become outgoing and sectors p2 become incoming. The uncoupled indices in splitting tree f₁ and fusion tree f₂ have levels (or depths) levels1 and levels2 respectively, which determines how indices braid. In particular, if i and j cross, $τ_{i,j}$ is applied if levels[i] < levels[j] and $τ_{j,i}^{-1}$ if levels[i] > levels[j]. This does not allow to encode the most general braid, but a general braid can be obtained by combining such operations.

TensorKit.permute — Method

permute(f₁::FusionTree{I}, f₂::FusionTree{I},
        p1::NTuple{N₁, Int}, p2::NTuple{N₂, Int}) where {I, N₁, N₂}
-> <:AbstractDict{Tuple{FusionTree{I, N₁}, FusionTree{I, N₂}}, <:Number}

Input is a double fusion tree that describes the fusion of a set of incoming uncoupled sectors to a set of outgoing uncoupled sectors, represented using the individual trees of outgoing (t1) and incoming sectors (t2) respectively (with identical coupled sector t1.coupled == t2.coupled). Computes new trees and corresponding coefficients obtained from repartitioning and permuting the tree such that sectors p1 become outgoing and sectors p2 become incoming.